hotfix: 퀴즈 api 배포 오류 수정
